+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 This page reports the main releases only and the main changes therein.
 
-## [v1.11.x] - 2025-11-29
+## [v1.11.x] - 2025-11-30
 **Unreleased changes**
 
 ### Fixed
+- fixed threshold parameter behavior un unsharp mask filter
 - prevent failure in case a single file is selected for parallel alignment
+- reverts to sequential pyramid focus stacking if temporary disk space is insufficient
+
+### Changed
+- optimized temporary disk cache for parallel pyramid focus stacking
 
 --------------------------